+++ tests/solver.py
@@ -10,7 +10,7 @@ def assert_satisfiable(f: Prop, msg: str = ""):
     assert not f.is_false_atom()
     if f.is_true_atom():
         return
-    base, naux, cnf = op.compute_cnf((f,))
+    base, naux, cnf = op.compute_cnf(f)
     with Solver(bootstrap_with=CNF(from_clauses=cnf)) as solver:
         assert solver.solve(), msg
 
@@ -18,7 +18,7 @@ def assert_unsatisfiable(f: Prop, msg: str = ""):
     assert not f.is_true_atom()
     if f.is_false_atom():
         return
-    base, naux, cnf = op.compute_cnf((f,))
+    base, naux, cnf = op.compute_cnf(f)
     with Solver(bootstrap_with=CNF(from_clauses=cnf)) as solver:
         assert not solver.solve(), msg